chat-model/requirements.txt
fastapi
uvicorn
pika
requests
llama_cpp_python
